Booting works on both linuxu and kvm
Basic program with init/destroy functions called are working.
main.c includes smoltcp.h, where C structs and extern functions implemented
in rust are defined. In main, the application will be written.

Using NoHashHasher because std::HashMap with default hasher was crashing the
unikernel, because of syscall 318 not being implemented (it was called because
of the unix/rand.rs file.

Signed-off-by: Florin Diaconescu <florin.diaconescu@stud.acs.upb.ro>