Added main.c and smoltcp.h
It provides a simple header for add_socket, add_socket_with_buffer (for custom
rx and tx buffer size) and initing/destroying the stack. The return value from
the Rust shim is good, and in case of adding a socket not yet supported (a dummy
TCP and UDP implementation is provided for now), Rust panics and the program
exits with the error message and a segmentation fault.

Signed-off-by: Florin Diaconescu <florin.diaconescu@stud.acs.upb.ro>