Loopback interface can be created
Using InterfaceBuilder and function build_interface() exposed in smoltcp.h
Changed device field from Stack structure from DeviceT to Option<DeviceT> to
take ownership when calling InterfaceBuilder::new. Now stack.device.take().unwrap()
is passed, as we first have to take ownership of the Option<DeviceT> and then
unwrap that value into the corresponding result:
-None -> the function panics
-Some(DeviceT) -> it returns the DeviceT parameter

Same change was made to neigh_cache field, to take ownership of the field
and to initialize it to None when calling Stack::new().

Signed-off-by: Florin Diaconescu <florin.diaconescu@stud.acs.upb.ro>