Added smoltcp_{listen,connect} functions
These are currently implemented for usage with any interface, however they were
tested only on Loopback device. First, listen is called on the server socket
(using a specified port) and then the client connects to the server endpoint
with the server IP (in Loopback's case 127.0.0.1) + port, and the client's own
port.

Signed-off-by: Florin Diaconescu <florin.diaconescu@stud.acs.upb.ro>