TCP client-server works on Loopback interface
The main.c code works in a while(1) loop. In each iteration, poll_interface()
is called and there is code for both the client and the server. First, the
server listens, the client connects to the server and when the client can_send()
a short buffer is sent (which is received by the server).

Signed-off-by: Florin Diaconescu <florin.diaconescu@stud.acs.upb.ro>