Refactoring + main cleanup
When getting a socket handle from the interface's SocketSet, we couldn't know
if the socket was an UDP/TCP one. Therefore, now our internal handle map has
a tuple (SocketHandle, SocketType) as its key, to call the get function
accordingly to the type. This is useful for functions that apply to both TCP
and UDP sockets (for example, bind only makes sense for UDP sockets and listen
only for TCP sockets).

The main.c was simplified, as we were using a simple transmit/receive with a
queue size equal to 1. Those functions will be moved to a different C file
that will get compiled when including smoltcp as an Unikraft external library,
to reduce the main program complexity for the user.

Signed-off-by: Florin Diaconescu <florin.diaconescu@stud.acs.upb.ro>