Simplify uk_recv code
Added a new structure, PacketInfo, which contains fields for an void *(the nb->data)
and an uint16_t (for the package length, nb->len). These are used for converting
in our shim layer between the raw uknetdev package to a raw Ethernet frame that
is understood by smoltcp. This frame can be printed using smoltcp::PrettyPrinter,
get modified and emited back to a frame understood by uknetdev.

Signed-off-by: Florin Diaconescu <florin.diaconescu@stud.acs.upb.ro>